requirements.txt
aioredis==1.3.1
argcomplete==1.11.1
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
autopep8==1.5.2
beautifulsoup4==4.9.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.1
cycler==0.10.0
daphne==2.5.0
decorator==4.4.1
distro==1.5.0
Django==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-fontawesome-5==1.0.18
django-nine==0.2.3
django-rq==2.3.1
django-templated-mail==1.1.1
django-websocket-redis==0.6.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
gevent==20.4.0
greenlet==0.4.15
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
isort==4.3.21
keyboard==0.13.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
msgpack==0.6.2
networkx==2.4
numpy==1.17.4
pandas==0.25.3
path==13.2.0
Pillow==7.1.1
pipx==0.15.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
rq==1.3.0
rstr==2.2.6
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.7
userpath==1.3.0
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.1.0
